Re-enable the tflite functions in final_pipeline
