--- inference_cpp/CMakeLists.txt
@@ -1,1 +1,0 @@
-
